Compartmentalized the velocity scheme and function: adflow from 2D notebook.
Function adflow contains code which a.) defines the grid b.) runs a loop over a
specified time interval, c.) produces plots of the results, and d.) writes
the solution to a file in text form.

Appended a list of arguments that should eventually be inputs of adflow.

Function u_simple contains code which defines the simple rotational velocity
scheme on a specified grid.