Wrote test function for dissolved and particulate Pa, and cleaned up adflow, u_simple, u_complex.
Removed some unecessary inputs from adflow. Added aforementioned input to velocity functions.

The test function tests that the output of soon to be written function "upwind" is the same as
the output from adflow before the time loop was removed. Upwind will contain the time loop, and take
many inputs in common with adflow. adflow will take then upwind function as an input, and eventually other
advections schemes as well.