Edited test functions to call upwind using adflow.
Sicne this is how the code will be run in the model, this is what needs to be checked. For some reason, when calling upwind with the inputs adflow gives it, but not inside adflow, the solution is a very poor match. I have a hunch it was human error, because thankfull when upwind is called by adflow, the solution is within ~1e-8 at the most from the solution to adflow before the time loop was removed and implemented in upwind.